chore(ethereum): update package exports
